requirements.txt
jinja2
aiofiles    
xmltodict
pandas
s3fs 
fsspec
pydantic

#for database
sqlalchemy
psycopg2

#for loading environment variables
python-dotenv